requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
colorama==0.4.6
pyfiglet==1.0.2
googlesearch-python==1.2.3
questionary==2.0.1
tqdm==4.66.1
tabulate==0.9.0
